Move Parser::and_or_list
